Servers/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.2.1
django-cors-headers==3.0.1
django-countries==5.3.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-import-export==1.2.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
et-xmlfile==1.0.1
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10.1
MarkupSafe==1.1.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.5.12
PyJWT==1.7.1
pytz==2019.1
PyYAML==3.13
requests==2.22.0
simplejson==3.16.0
sqlparse==0.3.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.2
virtualenv==16.6.0
xlrd==1.2.0
xlwt==1.3.0
